new file:   src/mpl_bitwise.c 	modified:   src/mpl_set.c 	new file:   src/mpl_to.c 	modified:   Makefile 	modified:   include/mpl.h 	deleted:    src/mpl_to_str.c

+++ src/mpl_set.c
@@ -8,7 +8,7 @@
 #include "mpl_common.h"
 
 void
-mpl_set_sint(mpl_int *a, long val)
+mpl_set_sint(mpl_int *a, int val)
 {
 	int i, oldtop;
 	_mpl_int_t *dp;
@@ -30,7 +30,7 @@ mpl_set_sint(mpl_int *a, long val)
 }
 
 void
-mpl_set_uint(mpl_int *a, unsigned long val)
+mpl_set_uint(mpl_int *a, unsigned int val)
 {
 	int i, oldtop;
 	_mpl_int_t *dp;
